support loading tar/parquet from local path
