+++ finetrainers/data/__init__.py
@@ -24,4 +24,3 @@ from .precomputation import (
     initialize_preprocessor,
 )
 from .sampler import ResolutionSampler
-from .utils import find_files
